feat: Better package visibility
This basically removes the exported maps and iotas for month names.
Instead, we have exported functions that return these values with proper
tests and error handling so that the logic is self contained within the
package. It also creates a nicer godoc representation - users do not
need to see the 200 line map of years to days in month slice.