fix: Avoid representing BS dates using time package to fix overflows (#24)
Fixes #23 

If a B.S. date (say जेठ 32) is represented as 02/32 (which was the case right now), toTime will (obviously) interpret it as Feb 32 which is an overflow by 32 - 29 = 3 days. So it's reinterpreted by the time package to be March 3 which is interpreted by nepcal to be असार 3.

This is now fixed by adding a separate struct BSDate.