bug: Fix tests failing on Go 1.13
The issue is an upstream one. Refer to https://github.com/golang/go/issues/31859 for all the details.
On a surfacial level, it breaks client code that calls `flag.Parse` in `init` functions. The solution,
albeit unsatisfactory, is to move the `flag.Parse` call to a regular function instead. This pollutes
the test output slightly but CI *should* pass now.